refactor(test): test 유틸리티 클래스 패키지 변경 - test 유틸리티 패키지 위치: bc.aggregate.testUtil -> bc.testUtil
