+++ src/test/java/com/simpleboard/board/board/domain/post/entity/PostDeleteTest.java
@@ -8,8 +8,8 @@ import com.simpleboard.board.board.domain.post.dto.CreateParams;
 import com.simpleboard.board.board.domain.post.dto.DeleteParams;
 import com.simpleboard.board.board.domain.post.exception.MemberPostPermissionException;
 import com.simpleboard.board.board.domain.post.exception.PostPasswordNotMatchException;
-import com.simpleboard.board.board.domain.post.testutil.PostCreateParamsBuilder;
-import com.simpleboard.board.board.domain.post.testutil.VisitorUtil;
+import com.simpleboard.board.board.domain.testUtil.PostCreateParamsBuilder;
+import com.simpleboard.board.board.domain.testUtil.VisitorUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
