+++ src/test/java/com/simpleboard/board/board/domain/post/entity/PostLikeTest.java
@@ -3,7 +3,7 @@ package com.simpleboard.board.board.domain.post.entity;
 import static org.assertj.core.api.Assertions.*;
 
 import com.simpleboard.board.board.domain.common.vo.Visitor;
-import com.simpleboard.board.board.domain.post.testutil.VisitorUtil;
+import com.simpleboard.board.board.domain.testUtil.VisitorUtil;
 import com.simpleboard.board.board.domain.post.vo.PostTypeEnum;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
