+++ src/test/java/com/simpleboard/board/board/domain/post/vo/LikesVOTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import com.simpleboard.board.board.domain.common.vo.Visitor;
 import com.simpleboard.board.board.domain.post.dto.LikeInfo;
 import com.simpleboard.board.board.domain.post.entity.PostLike;
-import com.simpleboard.board.board.domain.post.testutil.VisitorUtil;
+import com.simpleboard.board.board.domain.testUtil.VisitorUtil;
 import org.junit.jupiter.api.*;
 
 /**
